Rename SymbolTree to Term
